+++ projectroles/templatetags/projectroles_common_tags.py
@@ -1,11 +1,21 @@
 """Template tags provided by projectroles for use in other apps"""
 
+import mistune
+
 from django import template
 
 from projectroles.plugins import get_backend_api
 
 
 register = template.Library()
+
+
+@register.simple_tag
+def render_markdown(raw_markdown):
+    if type(raw_markdown) == bytes:
+        raw_markdown = raw_markdown.decode('utf-8')
+
+    return mistune.markdown(raw_markdown)
 
 
 @register.simple_tag